disable lint check for single quoted variables
Previously we weren't linting, there's a number of warning but there was
only one actual error reported when linting was enabled.  It was
erroring on a valid usage of a single quoted variable-like string.
To allow linting to succeed (and the warning to show themselves) I've
disable the check that is giving a false positive.