(MODULES-4973) rip out data in modules
this module is not able to find a value for `$keyserver` when it is
included in postgresql, as stated in the ticket. this changes the module
back to using the params.pp model. we can revisit data in modules once
LTS moves up from 4.7.0. for now, we need the compatibility.