(MODULES-10763) loglevel won't affect reports
The README suggests that setting a sufficiently high `loglevel` metaparameter
value will squelch the `Exec[apt_update]` resource's corrective changes from
agent reports. This is not true. Correct and clarify this statement to limit
its scope to log output, and link to MODULES-10763 for context on the exec
resource's false-positive reporting of corrective changes as a known issue.