MODULES-10956 remove redundant code in provider apt_key
The code in function `update_expired_key` is redundant to the logic in
the function `create`. The call to function `exist` will mark the key as
absent, in case it has expired and refres is set to true.

I debugged for hours why my keys weren't updated until I realized, that
the functionality is elsewhere.