(ISSUE-1036) Conditional `gnupg` include added to init.pp
Originally removed as it was causing `gnupg` to be installed in all OS when it wasn't needed, removing it seems to have caused a dependency cycle in the relevant Debian family OS for certain community members.
Adding the include back within a conditional statement to solve the issue while still preventing it from being included when unneeded.